pyproject.toml
[project]
name = "drf-template"
version = "0.1.0"
description = ""
authors = [
    {name = "kebato",email = "gulomovotabek444@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
